improved testing script

+++ fulltests.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 cd ~/redsat/test
+
+
+#make sure we are running the "actual" version of the judge
 rm -f juezejecutable
 (cd ~/redsat && make && cp juezejecutable test)
 
@@ -14,18 +17,34 @@ function run_problem {
 	f="../problems/${problem}/"
 	fproposal="${f}proposals/${proposal}/"
 	./juezejecutable "${f}"jp "${f}"jp2input "${f}"input2sat "${fproposal}"propuestasolucion "${f}"validator "${f}"format > std.out 2> std.err
-	#echo "diff answer answer.cor"
-	diff answer "${fproposal}answer.cor"
-	#echo "diff answer.long answer.long.cor"
-	diff answer.long "${fproposal}answer.long.cor"
-	#echo "diff std.out std.out.cor"
-	diff std.out "${fproposal}std.out.cor"
-	#echo "diff std.err std.err.cor"
-	diff std.err "${fproposal}std.err.cor"
+
+	#the test failed if either answer.long or std.out differs
+	if [[ $(diff -q answer.long ${fproposal}answer.long.cor) ]] || [[ $(diff -q std.out ${fproposal}std.out.cor) ]]; then
+		echo "===== FAILED TEST ====="
+		echo ">>>> diff answer answer.cor"
+		diff answer "${fproposal}answer.cor"
+		echo ">>>> diff answer.long answer.long.cor"
+		diff answer.long "${fproposal}answer.long.cor"
+		echo ">>>> diff std.out std.out.cor"
+		diff std.out "${fproposal}std.out.cor"
+		echo ">>>> diff std.err std.err.cor"
+		diff std.err "${fproposal}std.err.cor"
+
+		#execute again with debug flag, save output in test/dbg
+		./juezejecutable -d "${f}"jp "${f}"jp2input "${f}"input2sat "${fproposal}"propuestasolucion "${f}"validator "${f}"format > dbg 2> /dev/null
+
+		#dont continue running tests
+		exit
+	fi
+	
+	#if the test doesn't fail, don't print anything
 
 	rm -f answer answer.long std.out std.err
 }
 
+run_problem minitest correct
+run_problem "test" correct
+run_problem sudoku correct #does not contain cardinality constraints
 run_problem nreinas correct
 run_problem nreinas badred
 run_problem nreinas badrec
